added method to handle errors
