+++ sources/packages/web/src/services/StudentService.ts
@@ -1,6 +1,6 @@
 import ApiClient from "@/services/http/ApiClient";
-import { StudentProfile, SINValidations, ApiProcessError } from "@/types";
-import { useFormatters, useSnackBar } from "@/composables";
+import { StudentProfile, SINValidations } from "@/types";
+import { useFileUtils, useFormatters } from "@/composables";
 import {
   AESTFileUploadToStudentAPIInDTO,
   StudentFileDetailsAPIOutDTO,
@@ -15,7 +15,6 @@ import {
   UpdateDisabilityStatusAPIInDTO,
 } from "@/services/http/dto";
 import { AxiosResponse } from "axios";
-import { FILE_HAS_NOT_BEEN_SCANNED_YET, VIRUS_DETECTED } from "@/constants";
 
 export class StudentService {
   // Share Instance
@@ -152,14 +151,7 @@ export class StudentService {
         `student/files/${uniqueFileName}`,
       );
     } catch (error: unknown) {
-      if (
-        error instanceof ApiProcessError &&
-        (error.errorType === FILE_HAS_NOT_BEEN_SCANNED_YET ||
-          error.errorType === VIRUS_DETECTED)
-      ) {
-        const snackBar = useSnackBar();
-        snackBar.warn(error.message);
-      }
+      useFileUtils().handleFileApiProcessError(error);
       throw new Error(
         "There was an unexpected error while downloading the file.",
       );
